requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
faiss-cpu
sentence-transformers
torch
langchain
pandas
numpy
transformers